split out new data registry responsible for dataset overview, vs data manager resonsible for data downloads and zarrs
