+++ src/eo_api/data_manager/services/__init__.py
@@ -1,1 +1,1 @@
-from . import constants, datasets, downloader, utils
+from . import constants, downloader, utils
