+++ src/eo_api/data_registry/__init__.py
@@ -1,0 +1,1 @@
+from . import routes, services
