+++ src/eo_api/data_registry/services/__init__.py
@@ -1,0 +1,1 @@
+from . import datasets
