+++ src/eo_api/main.py
@@ -4,7 +4,7 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
 import eo_api.startup  # noqa: F401  # pyright: ignore[reportUnusedImport]
-from eo_api import data_accessor, data_manager, system
+from eo_api import data_accessor, data_manager, data_registry, system
 
 app = FastAPI()
 
@@ -17,5 +17,6 @@ app.add_middleware(
 )
 
 app.include_router(system.routes.router, tags=['System'])
-app.include_router(data_manager.routes.router, prefix='/datasets', tags=['Data manager'])
+app.include_router(data_registry.routes.router, prefix='/registry', tags=['Data registry'])
+app.include_router(data_manager.routes.router, prefix='/manage', tags=['Data manager'])
 app.include_router(data_accessor.routes.router, prefix='/retrieve', tags=['Data retrieval'])
